% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-competitions.R
\name{extract_competition_ids}
\alias{extract_competition_ids}
\title{Get the ids for a specific competition.}
\usage{
extract_competition_ids(response, competition)
}
\arguments{
\item{response}{The response returned by \code{get_competitions}.}

\item{competition}{Competition you want the ids for e.g. Premier League.}
}
\value{
List of both the competition id and the season ids. The competition id is a single number, the season ids are
a named character vector with id by season.
}
\description{
\code{extract_season_ids} gets the ids for a specific competition.
}
\details{
You can only get the season ids for a single competition at a time
}
