% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extract_season_ids}
\alias{extract_season_ids}
\title{Get the season ids for a specific competition.}
\usage{
extract_season_ids(response, competition)
}
\arguments{
\item{response}{The response returned by \code{get_competitions}.}

\item{competition}{Competition you want the ids from e.g. Premier League.}
}
\value{
Named character vector with id by season
}
\description{
\code{extract_season_ids} gets the season ids for a specific competition.
}
\details{
You can only get the season ids for a single competition at a time
}
