% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fixtures.R
\name{get_fixtures}
\alias{get_fixtures}
\title{Get played fixtures.}
\usage{
get_fixtures(competition_id, season_id)
}
\arguments{
\item{competition_id}{Can be obtained through the competitions resource.}

\item{season_id}{Can be obtained through the competitions resource.}
}
\value{
Response with information about each fixture (match id, attendance, ...)
}
\description{
\code{get_fixtures} returns info about fixtures played.
}
\details{
The request parameters are the page size and level of detail. These can't be changed since there are only a few
sensible values. Page size is always fixed at 1000 (we don't expect more than 1000 games in a single season).
}
\seealso{
\code{\link{get_premier_league}}
}
