% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import.R
\name{get_premier_league}
\alias{get_premier_league}
\title{Helper for functions making GET request to the Premier League API.}
\usage{
get_premier_league(resource, parameters, timeout_in_sec = 5,
  wait_in_sec = 1)
}
\arguments{
\item{resource}{type of resource you want to get}

\item{parameters}{list of parameters}

\item{timeout_in_sec}{after this amount of time a time-out error is thrown}

\item{wait_in_sec}{idle time in seconds before actual request (as a matter of decency it's common not to flood the server with requests)}
}
\value{
Response including status code, body, ...
}
\description{
\code{get_premier_league} is a helper for functions making GET request to the Premier League API.
}
\details{
Function not called directly.
}
