% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_premier_league}
\alias{get_premier_league}
\title{Create functions making GET request to the Premier League API.}
\usage{
get_premier_league(parameters)
}
\arguments{
\item{parameters}{List of parameters}
}
\value{
Function to make GET request to Premier League API.
}
\description{
\code{get_premier_league} creates functions making GET request to the Premier League API.
}
\details{
Function created will return the complete response body including info like headers and status code.
}
