% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import.R
\name{get_premier_league}
\alias{get_premier_league}
\title{Helper for functions making GET request to the Premier League API.}
\usage{
get_premier_league(resource, parameters, timeout_in_sec = 5)
}
\arguments{
\item{resource}{Type of resource you want to get}

\item{parameters}{List of parameters}

\item{timeout_in_sec}{After this amount of time a time-out error is thrown}
}
\value{
Response including status code, body, ...
}
\description{
\code{get_premier_league} is a helper for functions making GET request to the Premier League API.
}
\details{
Function not called directly.
}
